Further harden the VerifyDecryptKeyExchange tests
Running the test in a loop with the additional logging showed that the test can
fail, due to the nature of the randomness and underlying cryptographic algorithm.

By also checking that the return value isn't the pre-tamper we should reduce the
failure rate from "once every 3.5 months" (amortized current known rate) to
"once every hundred trillion years", give or take.