Fix 79 System.Data.Common.Test failures on ILC
These tests were null-referencing inside a IsReadOnly
property that looked up ReadOnlyAttributes in a collection
and invoked it without null-checking the lookup result.

The reason it's allowed to do that that is this
collection has some magic in it that lazily auto-populates
itself with the "default" value of the attribute
if the attribute itself isn't found on a lookup.
So as long a default exists, a lookup of this attribute
type never returns null.

Unfortunately, the "default" is discovered by reflecting
for a static field named "Default" on the attribute
type.