Make sure source isn't included in NETStandard.Library.NETFramework
Other framework packages (Microsoft.Private.*) are transport packages
and need to contain source and symbols in order to flow those up to the
actual framework packages that are built in core-setup.

We put the source files in the sources path in our other framework pacakges
because we need to workaround issues with myget assuming a symbols package
when it sees a src folder.

NETStandard.Library.NETFramework is not a transport package so we shouldn't
put its sources in the sources folder.  Instead they should be in the src
folder so that pack splits them out into the symbols package.