Fix Windows X509Certificate2.ToString(true) for ephemeral private keys
AppendPrivateKeyInfo handled not being able to resolve the container details
for a private key, but an ephemeral private key caused a NullReferenceException
instead of the caught CryptographicException.

The comments in the code suggest this worked properly at one time, but there
was no test guaranteeing functionality.  This change handles the null return from
GetPrivateKeyCsp, and adds some tests to ensure this doesn't break again.