Make all X509Store.Open exceptions be CryptographicException. (#19844)
The platform limitations are now
CryptographicException(PlatformNotSupportedException) instead of PNSE.

This should restore some user expectation around the exception model, given
that the call to Open can be delayed from the call to an X509Store .ctor.

This also softens the exceptions from the Disallowed store on Linux to let empty
reads succeed, only writes and pre-populated data will fail.

Updated the X509Store table in the xplat crypto doc.