Fix RevocationMode and RevocationFlag errors on macOS
For certificates which have CRL, but not OCSP, the CRL doesn't seem to be
checked, then Apple is silent about the matter.  By setting
kSecRevocationRequirePositiveResponse we get notified "RevocationResponseRequired".
That flag should be mapped to X509ChainStatusFlags.RevocationStatusUnknown.

When doing NoCheck, the OS was opportunistically checking revocation,
possibly from a cache.  So make sure that doesn't show up.

Also, filter down the results based on the RevocationFlag value.