Move reference expansion to a target and handle simple name conflicts
We cannot let simple name references that overlap with our assemblies reach
RAR because RAR will choose the files in the targeting pack over the ones we
add to the reference item, due to the default AssemblySearchPaths.

We could change AssemblySearchPaths to consider all the NuGetPackage paths
before `{TargetFrameworkDirectory}` but that has broader implications for
compatibility and performance of RAR.

I chose to follow a similar pattern to what the SDK does already and took the
opportunity to move the reference item setting into the target as well, to more
closely match what the SDK is doing.