Add ShellEx support for Windows & fix issues (#20216)
* Add ShellEx support for Windows & fix issues

Adds ShellExecute support for Windows (outside of UAP). Unlike desktop
UseShellExecute is not the default.

In the process of ding this I ran across two issues:

1- The owned process handle would be disposed after hitting members on
Process. Getting the ProcessName, for example, would cause a closed
handle exception on the next call that used the handle.

2- The MainWindowTitle import wasn't set to the right charset. In addition
the allocated buffer size wasn't correct. Fixed and added debug checks
for failed api calls.

* Address feedback

- Add usings in tests
- Handle missing API

* Move to new using pattern

- Remove newly dead string
- Other feedback

* More feedback changes