Workaround pooling issue with managed HttpListener and Windows ClientWebSocket
The ClientWebSocket on Windows appears to be using connection pooling.  If it connects to the managed HttpListener, which subsequently closes the connection, and that closing races with the client trying to re-use the connection, it appears to cause the client request to fail.  Rather than retry, as a workaround we simply change the tests to try to create each temporary HttpListener on a different port than the last one, and only reuse a port once we've exhausted all of them.  This then avoids any issues related to connection pooling, as each test will be connecting to a listener on a different port.