Get rid of assorted CoreFx test on ILC noise.
The recent change to unblock all serializable types
for Reflection (even if the types themselves are
internal) woke up the XUnit ToString() problem
again.

Now, reflecting on ToString() on these types
are throwing MME's rather than bleeding through
the prevaricator Type that pretends to have no
members to the public base type.

This is most irritating but there's not much for
it except to rd.xml these internal types
directly.