Set PFX temporary keychains to have a lock policy of never.
The default lock policy on a new keychain is to lock on sleep and after 5 minutes
of being idle.  This means that if more than 5 minutes pass from usages of a
certificate private key (where the cert was opened from a PFX) the user gets
prompted for a password for the temporary keychain.  Since that keychain has
a randomly generated password it's not really possible for the user to accomplish
this successfully.

No tests cover this scenario, since they require more than 5 minutes of idle time.