Workaround for recent CMake bug w.r.t. CF guard (#20831)
Recent versions of CMake generates the <LinkControlFlowGuard> node
under the <Link> node when it finds /guard:cf linker option in the
linker options. But that is not correct. It should generate it under
the <PropertyGroup> node so that the Microsoft.CppCommon.targets would
then generate <ControlFlowGuard> node under the <Link>.

To workaround the problem, change the casing of /guard:cf to uppercase.
CMake then doesn't detect the option as something known to it and passes
the option to the linker as additional option, which fixes the problem.