Bring SerializationFieldInfo back to BinaryFormatter in corefx (#20938)
We initially left SerializationFieldInfo out of corefx as it wasn't necessary given previous goals around BinaryFormatter; but now it is.  I've ported the code almost as-is from desktop, with some minor style tweaks, and a few necessary changes, e.g. using FieldInfo instead of RuntimeFieldInfo, as the latter isn't accessible.  I've not done anything to minimize allocations / improve perf, etc., as that can be done in the future, and the results of all of this gets cached.

Changing this required regenerating a few of the serialization blobs in the tests.  I also added a test for a case we didn't previously cover, a [Serializable] type derived from a non-serializable type.