Step 1 of fix for disabled System.Dynamic.Runtime test on ILC
(https://github.com/dotnet/corefx/issues/19895)

Make Microsoft.CSharp use the the real HasMetadataDefinitionAs()
api if it exists on the running framework. Since this library
is built on netstandard, it still needs to fall back on its
emulation on older frameworks.

Actual removal of the ActiveIssue will have to wait until
the new System.Runtime reference assembly propagates to
the Project N tree and the new api becomes reflectable.