Change PathList members from public to private (#21142)
PathList was made public to enable binary serialization of
CookieContainer. PathList is not exposed in the reference assembly as it
is an implementation detail and should not be used directly by external
callers. Out of an abundance of caution, make the public members
internal to prevent anyone from inadvertently taking a dependency on
them, which gives us some flexibility if we want to make changes in the
future.