Disable User and Machine environment api tests inside appcontainers. (#21268)
* Disable User and Machine environment api tests inside UAP/UapAot

We're not going to get exemptions for these inside appcontainers
so we'll change them to behave as they do on Unix (present
an empty reg key that eats writes requests.)

Since this requires a coordinated change against three repos,
the first step is to disable the tests entirely until
we get the other parts done.

* Replace IsWinRT || IsNetNative with IsUap

Also, CoreCLR and CoreRT won the race to change
their behavior so this is now the final commit
to fix the tests. Removing the probing check
altogether.