Re-enable some System.Net.Mail tests for UAPAOT
With the recent implementation of System.Net.NetworkInformation for UWP,
we can re-enable some tests.

Also, the latest plan of record for serialization support in .NET Core
is such that serializing these exception types are not supported. So,
removing the test for that.  FYI, the actual implementation of the
serialization overrides in the SmtpException class are already no-ops in
the code and simply call the base method.

Fixes #19604
Fixes #19585