Add ClientWebSocket client certificate test (#21466)
Adding new test case that pins the UWP shortcoming tracked by issue #21393. This test case requires special setup that is checked inside its ConditionalFact attribute.

I've confirmed that it consistently passes in .NET Framework test runs and
fails in UWP and .NET Core test runs (as expected). I'll enable it for UWP as part of the #21393 fix.

Contributes to #21393