Fix use of Cookie.InternalSetName
The System.Net.Primitives UAPAOT test run was failing in
SetCookies_InvalidData_Throws. The reflection wasn't working as
expected.  However, this same CookieParser.cs file is included in the
System.Net.HttpListener library.  And that reflection was working.

It seems that reflecting on a type (Cookie) from within the same library
(System.Net.Primitives) doesn't work from another type in that same
library (CookieParser).

It turns out that we can optimize this anyways and not use reflection
when the file is included directly in System.Net.Primitives.

Fixes #20482