Fix some TLS related HttpClientHandler properties for UAP
Due to platform differences on UAP and the current use of WinRT APIs to
implement that HTTP stack, a few properties won't have the same defaults
or behavior.

CheckCertificateRevocationList default value is true for UAP and can't
be changed. However, on UAP, revocation checking that results in an
indeterminate status (i.e. revocation server offline) is treated the
same as a valid result. This is the behavior of the underlying WinRT
APIs.

SslProtocols default value is SslProtocols.None which is the same as the
rest of .NET Core. 'None' means to use the system default. However, changes
to this property value will be ignored.

Similar to the decision made in PR #21403, we don't throw PNSE for these
differences. Instead we just no-op on the property setter.

The behavior differences will be documented in:
https://github.com/dotnet/corefx/wiki/ApiCompat

Contributes to #18116.