Move Entries from Library Rd.xml to Test Rd.xml.
XmlSerializer.Mode iss marked as required by the System.Private.Xml.rd.xml.
Unfortunately, it pulls in the XmlSerializer type itself (has to), which
in turn pulls in a whole lot of code through its static ctor dependencies.
This caused some size-on-disk regressions.

As the property is used by our test project only, the fix is to move the
entry from the library rd.xml to the test rd.xml.

Fix #21667