Make fields on HttpRequest/ResponseHeaders pay-for-play
There are currently 9 reference fields on HttpResponseHeaders and 5 reference fields on HttpRequestHeaders that are used to cache lazily-instantiated collections when accessing strongly-typed properties on the headers objets.  That's 14*8==112 bytes of fields allocated per request, even if not used.  This commit lifts those fields into a lazily-allocated array; this makes them a bit more expensive when they're used, but reduces allocation when they're not.