Prevent SslProvider exceptions from being thrown to native code
CurlHandler.SslProvider register the VerifyCertChain method to be invoked by OpenSSL in order to validate a cert chain.  If the validation code throws an exception from certain places, e.g. in response to OpenSSL reporting unknown cipher suites, that exception is currently allowed to propagate out of the function back across the interop boundary, which is unsupported and results in bad behaviors.  The fix is to widen the existing catch block to catch all such exceptions and marshal them to the associated request appropriately.