Fix segfault on chain build on macOS High Sierra (#21838)
When building a cert chain for self-issued certificates on High Sierra,
Security.framework gets into a bad state.  One minute after entering
SecTrustEvaluate the function exits with the trust result set to Invalid and
the return value (an OSStatus) set to success.

Continuing to interact with that chain results in more slowdowns and/or a
null-dereference triggering a segfault.

Hopefully the underlying issue is fixed before 10.13 RTM, but for now we can
be less destructive by throwing an exception instead of letting the native code
terminate the process.