Iterate TypeArray's inner array instead of TypeArray when safe.
Rewrite enumerations of TypeArrays to instead enumerate on the inner
array, allowing bounds check eliminations and removing some indirection.

This could be done in more places, but for now conservatively only done
where the "translation" is direct, so more easily confirmed as safely
equivalent to the existing code, and no local holds the array and no
slip-up could overwrite an element.