Remove BindExplicitConversionRequireOneArgument Error
Triggered if RuntimeBinder.BindExplicitConversion is called with any
number of arguments other than 1, but this is called by
CSharpConvertBinder.DispatchPayload which takes its arguments from the
args and parameters passed to RuntimeBinder.BindCore which is called by
RuntimeBinder.Bind which is called by BinderHelper.Bind which is called
by CSharpConvertBinder.FallbackConvert with a single argument, so any
other number of arguments cannot happen.

Replace with assertion.