Replace CSharpArgumentInfo.IsByRef with CSharpArgumentInfo.IsByRefOrOut
Microsoft.CSharp.RuntimeBinder.CSharpArgumentInfo has an internal
IsByRef method that is only used in conjunction with IsOut. Since these
are both flag tests, it can be replaced with a single IsByRefOrOut
test.

The same flag test is done inside RuntimeBinder, so replace that with a
call to this new method.

The Flags property is now no longer used from outside the class, so
make it private.