Remove unnecessary char[] allocation from StreamReader (#22147)
When StreamReader detects a different encoding than what it was initially told or defaulted to, it allocates a new char[] to hold the read chars.  But it really only needs to do this if the previously allocated char[] was too small, and since the default encoding is UTF8 and very few encodings have a larger byte:maxchar ratio than UTF8, we can avoid the allocation most of the time simply by checking to see if the existing array is large enough.