Simplify EmitExpressionAndBranch method.
EmitExpressionAndBranch is called from non-lifted non-method AndAlso
and OrElse, from Conditional and from Switch's use of Equal, and from
calls which originate with itself, all of which mean the node's type is
boolean. Removing the test for node.Type == typeof(bool) allows for the
rest of the method to be simplified.

Also include tests covering the EmitBranchNot case.
Contributes to #11409