Don't double-cast in nullable-to-nullable coalesce expressions.
If a coalesce of a nullable has the same type as the expression (e.g.
x ?? y if both x and y are int?) then not looking for this case means
if the lhs isn't null it's cast to non-nullable then back to nullable
again.

Instead catch this case and just use the left hand value as the result.