Make ManagedHandler ConnectionPoolingTest clean itself up
ManagedHandler_HttpClientHandler_ConnectionPooling_Test didn't assert
IDisposable (though did have a Dispose method), so it didn't clear the
environment variable that says to use the prototype managed handler in tests.
This means that any tests which should still be using the curl provider end up
using the managed provider until another managed test runs.

So now it's disposable, and a failure we were expecting, but not seeing, on
macOS 10.13 is asserting itself again.