Include base interfaces on lookup when dynamic is an argument to a call
When a variable is statically typed to an interface type that
implements another interface and an inherited method or indexer is
called with a dynamic argument, then the method or indexer will not be
found by the binder.

Change both the initial adding of methods to the table, and the
subsequent lookup, to fix this.

Fixes #14752