Include System.Object as base of interfaces on lookup when dynamic is an argument to a call
When a variable is statically typed to an interface and the Equals
method inherited from object is called with a dynamic argument, then
the method will not be found by the binder.

Change both the initial adding of methods to the table, and the
subsequent lookup, to fix this.

Fixes #22300