Optimise CreateInheritanceHierarchyList
Types are inserted to the front of a list, making n - 1 copies of n - 1
item (as well as any copies require), to ensure they are processed in the correct order.

Instead, add them to the end, and then iterate through the results in reverse.

Also create the list with the final capacity creating the list for an
interface type, as the capacity can be known.