Compile Boolean != as ^ in System.Linq.Expressions (#21806)
* Compile Boolean != as ^ in System.Linq.Expressions

Boolean and lifted Boolean != is identical to ^ in results.

In the case of the interpreter this allows for one type, and its
memoising field, to be removed, so do so.

In the case of the compiler this allows for a slightly shorter CIL
output. Detecting this adds to the compilation complexity, but also
inlining EmitUnliftedEquality into EmitUnliftedBinaryOp can remove a
branch on two tests so the nett result is simpler.

* Use single call with ?: inside more often than different calls on if.

Slightly smaller compiled size.