Microsoft.CSharp: Catch passing non-type as type for bound static call
If something other than a type is passed as the second argument (first
after the CallSite) then catch this and throw an explanatory
ArgumentException, rather than a RuntimeBinderInternalCompilerException
describing it as "unexpected".

Also remove assertion this cannot happen, as it clearly can.

Conversely, replace check that would throw
RuntimeBinderInternalCompilerException for this case in
CreateCallingObjectForCall with just an assertion, as the above check
will have happened before this point could be reached.

Fixes #22331