Simplify some S.L.Expressions branch compilation methods (#22166)
* Simplify EmitExpressionAndBranch method.

EmitExpressionAndBranch is called from non-lifted non-method AndAlso
and OrElse, from Conditional and from Switch's use of Equal, and from
calls which originate with itself, all of which mean the node's type is
boolean. Removing the test for node.Type == typeof(bool) allows for the
rest of the method to be simplified.

Also include tests covering the EmitBranchNot case.
Contributes to #11409

* Reduce if-else ladders in EmitAndAlsoBinaryExpression and EmitOrElseBinaryExpression

Handle all cases in 2 tests, instead of between 2 and 4.

* Simplify IL produced for EmitMethodAndAlso/OrElse

These methods dup the lhs, then if not short-circuiting store it, then
obtain the rhs and store than, the retrieve lhs and rhs and call
the method.

Instead dup the lhs and if not short-circuiting leave it on the stack,
obtain the rhs and call the method.