Remove Microsoft.CSharp paths for collection initialization
There's no way to express this to the dynamic binder (any use of
dynamic in an initializer is turned into Add calls at the static
compilation stage), so BindMethodGroupToArgumentsCore is only ever
called with bindingCollectionAdd as false).

Remove this parameter and those paths that depend on it being true.

Entails the deletion of ERR_BadArgTypesForCollectionAdd and
ERR_InitializerAddHasParamModifiers, contributes to #22470